NEW Ortho_Stats.df_indep_sampler method
Needed for statistical significance of DMF and similarity of taxonomical profiles.
This is the second attempt. It samples gene names independently, then merges
them on the index and drops NaN (so only existing pairs are left). Finally,
names pairs are merged with the DataFrame passed to Ortho_Stats class, creating
a subset.

Problems : e-value still larger/equal to the result. Wrong S.cerevisiae proteome
           might be the cause (taken from the InParanoid reference, not UniProt)
Important: unnecessary for the rest of the methods.
Misc     : Ortho_Stats.simple_profiles_scorer __doc__ amendment